Bluetooth: Host: Reassemble extended advertising reports
The host reassembles fragmented advertising reports from the controller.

Fragmented advertising reports from different advertisers may be
interleaved. If fragmented advertising reports from multiple advertisers
are received, only the first one is kept and reassembled. The rest are
discarded until the first one is complete.

If a complete advertising report is not reveived from an advertiser
within the timeout, the data is discarded and another advertiser is
allowed to be reassembled.